requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.2
Flask-Dance==6.0.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
oauth2==1.9.0.post1
oauthlib==3.2.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
urllib3==1.26.10
URLObject==2.4.3
Werkzeug==2.1.2
zipp==3.8.0
